requirements.txt
selenium
prompt_toolkit
rich
openai-whisper
python-telegram-bot
munch
openai
json5
textual
textual-dev
httpx
httpx[socks]
beautifulsoup4
